requirements.txt
Flask
Flask-Cors
firebase-admin
google-generativeai
python-dotenv
google-cloud-aiplatform
sentence-transformers
google-cloud-storage
sendgrid
